Backend/medplus/requirements.txt
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
distlib==0.3.7

filelock==3.12.2
idna==3.4
platformdirs==3.10.0
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.8.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
virtualenv==20.24.2
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
channels==4.0.0
decouple==0.0.7
Django==4.2.3
django-cors-headers==4.2.0
django-environ==0.10.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
frozenlist==1.4.0
multidict==6.0.4
phonenumbers==8.13.19
Pillow==10.0.0
python-decouple==3.8
yarl==1.9.2

